+++ sports-court-designer.php
@@ -3,7 +3,7 @@
  * Plugin Name: Sports Court Designer
  * Plugin URI: https://github.com/HaykSaakian/wp-court-designer
  * Description: Interactive sports court designer for tennis, basketball, and pickleball courts with customizable colors
- * Version: 1.5.1
+ * Version: 1.6.0
  * Author: Hayk Saakian
  * Author URI: https://github.com/HaykSaakian
  * License: GPL v2 or later
@@ -15,7 +15,7 @@ if (!defined('ABSPATH')) {
     exit;
 }
 
-define('COURT_DESIGNER_VERSION', '1.5.1');
+define('COURT_DESIGNER_VERSION', '1.6.0');
 define('COURT_DESIGNER_URL', plugin_dir_url(__FILE__));
 define('COURT_DESIGNER_PATH', plugin_dir_path(__FILE__));
 
@@ -128,7 +128,7 @@ class CourtDesigner {
         $valid_types = array(
             'tennis', 'basketball', 'pickleball',
             'tennis-1pb', 'tennis-2pb', 'tennis-4pb',
-            '2pb-tennis'
+            '2pb-tennis', 'pb-halfbasketball'
         );
         
         if (!in_array($atts['type'], $valid_types)) {